wip move map and set instance to alleycats
